requirements_dev.txt
pytest==5.3.5
sphinx==3.1.2
sphinxcontrib-napoleon
sphinx-autodoc-typehints
sphinx_rtd_theme
numpydoc
line-profiler==3.0.2
memory-profiler==0.57.0